+++ R/difNLR-package.R
@@ -48,8 +48,8 @@
 #' @details
 #' Package: difNLR\cr
 #' Type: Package\cr
-#' Version: 1.3.7-1\cr
-#' Date: 2021-02-16\cr
+#' Version: 1.3.7-2\cr
+#' Date: 2021-10-05\cr
 #' Depends: R (>= 3.1)\cr
 #' Imports: CTT, ggplot2 (>= 2.2.1), grDevices, methods, msm, nnet, plyr, reshape2, stats, VGAM\cr
 #' Suggests: ShinyItemAnalysis\cr